Prepare the projects data in FR and DE to be included in the app as .rds files

Prepare the projects data in FR and DE to be included in the app as .rds files

prepare_app_data(
  name_raw_file = "PGV.xlsx",
  pkg_dir = system.file(package = "exploratorium"),
  dic_variables = NULL,
  dic_cantons = NULL,
  cantons_sf = NULL
)

Arguments

name_raw_file

Character. Name of the raw data file.

pkg_dir

Character. Path to the package (must contain a data-raw folder).

dic_variables

Tibble. Variables dictionaries. Mainly used for examples and unit testing purpose.

dic_cantons

Tibble. Canton dictionary. Mainly used for examples and unit testing purpose.

cantons_sf

Sf data. Cantons geometry. Mainly used for examples and unit testing purpose.

Value

A list with the projects data in FR and DE.

Examples

# Load the toy datasets
data("toy_data_pgv")
data("toy_dic_variables")
data("toy_dic_cantons")
data("toy_cantons_sf")

# Create a temp folder with data-projects-raw and ata-projects subfolder
my_temp_dir <- tempfile("test-prepare-data")
dir.create(my_temp_dir)
dir.create(file.path(my_temp_dir, "data-projects-raw"))
dir.create(file.path(my_temp_dir, "data-projects"))

# Save the toy PGV file inside
writexl::write_xlsx(
  toy_data_pgv, 
  file.path(
    my_temp_dir, 
    "data-projects-raw", 
    "toy_PGV.xlsx"
  )
)

# Prepare the data
prepare_app_data(
  name_raw_file = "toy_PGV.xlsx",
  pkg_dir = my_temp_dir, 
  dic_variables = toy_dic_variables, 
  dic_cantons = toy_dic_cantons,
  cantons_sf = toy_cantons_sf
)
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Import the raw data
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Add the columns
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Clean the data
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Consolidate topic data
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Add the coordinates of the main organization
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> Passing 5 addresses to the Nominatim single address geocoder
#> Query completed in: 5 seconds
#> → Add the canton of the main organization
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Detect the id of the cantons influenced
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Add the number of cantons influenced
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Add the proportion of the budget
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Translate the data
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> ℹ Using "','" as decimal and "'.'" as grouping mark. Use `read_delim()` for more control.
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> → Save the data
#> ℹ Prepare projects_fr.rds and projects_de.rds
#> projects_fr.rds has been updated at 2024-05-13 10:07:26.401644
#> projects_de.rds has been updated at 2024-05-13 10:07:26.401644
#> ✔ Prepare projects_fr.rds and projects_de.rds ... done
#> 

# Delete the tempdir
unlink(my_temp_dir, recursive = TRUE)